Apply 2.13.0 release

+++ evkmimxrt595/driver_examples/powerquad/benchmark_sw_math/powerquad_benchmark_sw_math.c
@@ -240,7 +240,7 @@ static void arm_sin_f32Test(void)
 
     PRINTF("%s: %d ms\r\n", __func__, TEST_GetTime() - oldTime);
 
-    EXAMPLE_ASSERT_TRUE(fabs(sinRef - Result) < 0.00001);
+    EXAMPLE_ASSERT_TRUE(fabs((double)(sinRef - Result)) < 0.00001);
 }
 
 /* Q15 cos */
@@ -299,5 +299,5 @@ static void arm_cos_f32Test(void)
 
     PRINTF("%s: %d ms\r\n", __func__, TEST_GetTime() - oldTime);
 
-    EXAMPLE_ASSERT_TRUE(fabs(cosRef - Result) < 0.00001);
+    EXAMPLE_ASSERT_TRUE(fabs((double)(cosRef - Result)) < 0.00001);
 }
