+++ evkmimxrt595/driver_examples/powerquad/cmsis/powerquad_cmsis.c
@@ -237,7 +237,7 @@ static void arm_sin_f32Example(void)
 
     Result = arm_sin_f32(input);
 
-    EXAMPLE_ASSERT_TRUE(fabs(sinRef - Result) < 0.00001);
+    EXAMPLE_ASSERT_TRUE(fabs((double)(sinRef - Result)) < 0.00001);
 }
 
 /* Q15 cos */
@@ -278,7 +278,7 @@ static void arm_cos_f32Example(void)
 
     Result = arm_cos_f32(input);
 
-    EXAMPLE_ASSERT_TRUE(fabs(cosRef - Result) < 0.00001);
+    EXAMPLE_ASSERT_TRUE(fabs((double)(cosRef - Result)) < 0.00001);
 }
 
 /* Q15 Matrix Addition */
@@ -680,7 +680,7 @@ static void arm_mat_inverse_f32Example(void)
 
     for (uint32_t i = 0; i < 4; i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(inverseRef[i] - inverseMatrixR.pData[i]) < 0.00001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(inverseRef[i] - inverseMatrixR.pData[i])) < 0.00001);
     }
 }
 
@@ -1037,7 +1037,7 @@ static void arm_fir_f32Example(void)
 
     for (i = 0; i < ARRAY_SIZE(FIRRef); i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(FIRRef[i] - FIRResult[i]) < 0.0001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(FIRRef[i] - FIRResult[i])) < 0.0001);
     }
 
     /* Incremental method. */
@@ -1049,7 +1049,7 @@ static void arm_fir_f32Example(void)
 
     for (i = 0; i < ARRAY_SIZE(FIRRef); i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(FIRRef[i] - FIRResult[i]) < 0.0001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(FIRRef[i] - FIRResult[i])) < 0.0001);
     }
 }
 
@@ -1153,7 +1153,7 @@ static void arm_conv_f32Example(void)
 
     for (uint32_t i = 0; i < ARRAY_SIZE(ref); i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(ref[i] - result[i]) < 0.00001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(ref[i] - result[i])) < 0.00001);
     }
 }
 
@@ -1257,7 +1257,7 @@ static void arm_correlate_f32Example(void)
 
     for (uint32_t i = 0; i < ARRAY_SIZE(ref); i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(ref[i] - result[i]) < 0.00001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(ref[i] - result[i])) < 0.00001);
     }
 }
